fix(wdtt): починены 7 красных тестов пакета
Тесты приехали красными с 998d76d9, зелёными не были нигде.

- 4×TestApplyServerAccess*: applyServerAccess звал ensureServerWgClientRoute
  (реальный /opt/sbin/ip) и на NDMS-пути. После F2 адрес OpkgTun = 10.66.0.1/16
  (пул целиком), connected-маршрут уже покрывает клиентов — ручной маршрут
  оставлен только для legacy wdtt0/raw (usesNDMSAccess()==false).
- TestPrepareNDMSOpkgTunDoesNotSetAddress: assert искал "create OpkgTun17
  private", а fakeOpkgCommands пишет "create <name> <desc> <level>" —
  несовпадающий формат, тест поправлен под реальную запись.
- TestEnsureWdttIngressRefs: assert был инвертирован — "must preserve
  unrelated refs" проверялось как containsString(...) вместо !containsString(...),
  из-за чего тест требовал УДАЛЕНИЯ чужого ref вместо его сохранения.
- TestApplyRawServerPolicyWithGetterNonLinux: имя и комментарий теста говорят
  про non-linux стаб (server_raw_policy_other.go), но компилировался и гонялся
  в linux-сборке, дёргая реальный /opt/sbin/iptables. Вынесен в отдельный файл
  с //go:build !linux — по образцу остальных linux/other пар пакета.

Довесок из ревью Task 5: TestWdttNetfilterHookScriptShellSyntax гонял sh -n
без RawPolicyMark в spec — mangle-guard блок не проверялся. Добавлен
RawPolicyMark в spec теста.
