+++ internal/wdtt/entware_route_linux.go
@@ -10,8 +10,10 @@ import (
 )
 
 // ensureWgClientRoute adds a connected route for qWDTT monolith client pool
-// (10.66.0.0/16) via the kernel WG TUN. NDMS OpkgTun often configures only
-// 10.66.66.1/24, so replies to 10.66.0.x would otherwise leave via WAN.
+// (10.66.0.0/16) via the kernel WG TUN. Только для legacy wdtt0/raw: адрес там
+// 10.66.66.1/24, поэтому реплаи в 10.66.0.x иначе ушли бы через WAN. NDMS
+// OpkgTun сам получает 10.66.0.1/16 (PR #697, F2) — connected-маршрут /16 уже
+// покрывает пул, вызывающие на этом пути cfg.ensureServerWgClientRoute не зовут.
 func ensureWgClientRoute(ctx context.Context, iface, cidr string) error {
 	iface = strings.TrimSpace(iface)
 	cidr = strings.TrimSpace(cidr)
