+++ internal/wdtt/server_access_raw_test.go
@@ -32,13 +32,3 @@ func TestApplyRawServerPolicyWithoutGetter(t *testing.T) {
 		t.Fatalf("applyRawServerPolicy: %v", err)
 	}
 }
-
-func TestApplyRawServerPolicyWithGetterNonLinux(t *testing.T) {
-	svc := NewService(t.TempDir(), t.TempDir(), "", "")
-	svc.SetPolicyMarkGetter(&fakePolicyMarkGetter{mark: "0xffffaaa"})
-	cfg := ndmsServerConfig()
-	cfg.Policy = "Policy3"
-	if _, err := svc.applyRawServerPolicy(context.Background(), "srv1", cfg); err != nil {
-		t.Fatalf("applyRawServerPolicy: %v", err)
-	}
-}
